Fix some warnings
svn path=/trunk/; revision=24361

+++ reactos/tools/mkhive/cmi.c
@@ -31,7 +31,7 @@ CmpAllocate(
 	IN SIZE_T Size,
 	IN BOOLEAN Paged)
 {
-	return (PVOID) malloc(Size);
+	return (PVOID) malloc((size_t)Size);
 }
 
 static VOID
